fix: 자료형 변경

+++ tingting/src/main/java/com/alsif/tingting/book/repository/TicketSeatRepository.java
@@ -12,7 +12,7 @@ import com.alsif.tingting.concert.dto.concerthall.SeatBaseDto;
 import com.alsif.tingting.concert.entity.ConcertSeatInfo;
 
 @Repository
-public interface TicketSeatRepository extends JpaRepository<TicketSeat, Long> {
+public interface TicketSeatRepository extends JpaRepository<TicketSeat, Integer> {
 
 	@Query(
 		"SELECT NEW com.alsif.tingting.concert.dto.concerthall.SeatBaseDto(ts.ticket.seq, chs.section, chs.seat, g.price, g.name) "
@@ -21,11 +21,11 @@ public interface TicketSeatRepository extends JpaRepository<TicketSeat, Long> {
 			+ "    JOIN ConcertHallSeat chs ON (csi.concertHallSeat.seq = chs.seq) "
 			+ "    JOIN Grade g ON (csi.grade.seq = g.seq) "
 			+ "WHERE ts.ticket.seq IN (:ticketSeqs)")
-	List<SeatBaseDto> findAllPriceByTicketSeq(@Param("ticketSeqs") List<Long> ticketSeqs);
+	List<SeatBaseDto> findAllPriceByTicketSeq(@Param("ticketSeqs") List<Integer> ticketSeqs);
 
 	@Query("SELECT csi "
 		+ "FROM TicketSeat ts "
 		+ "JOIN ConcertSeatInfo csi ON (ts.concertSeatInfo.seq = csi.seq) "
 		+ "WHERE ts.ticket.seq = :ticketSeq")
-	List<ConcertSeatInfo> findAllByTicketSeq(@Param("ticketSeq") Long ticketSeq);
+	List<ConcertSeatInfo> findAllByTicketSeq(@Param("ticketSeq") Integer ticketSeq);
 }
