+++ tingting/src/main/java/com/alsif/tingting/concert/dto/concerthall/SeatBaseDto.java
@@ -16,8 +16,8 @@ import lombok.ToString;
 @Schema(description = "좌석 정보")
 public class SeatBaseDto {
 
-	@Schema(description = "예매 PK", type = "Long", example = "1768")
-	private Long ticketSeq;
+	@Schema(description = "예매 PK", type = "Integer", example = "1768")
+	private Integer ticketSeq;
 
 	@Schema(description = "좌석 코드 대분류", type = "String", example = "A")
 	private String section;
@@ -25,8 +25,8 @@ public class SeatBaseDto {
 	@Schema(description = "좌석 코드 소분류", type = "String", example = "D1")
 	private String seat;
 
-	@Schema(description = "좌석 가격", type = "Long", example = "77000")
-	private Long price;
+	@Schema(description = "좌석 가격", type = "Integer", example = "77000")
+	private Integer price;
 
 	@Schema(description = "좌석 등급", type = "String", example = "VIP")
 	private String name;
